@@ -231,7 +231,7 @@ class _ExpansionAlbumTileState extends State<ExpansionAlbumTile> {
                           child: Text(
                             appStrings.albumCount(widget.album.nbCategories),
                             style: TextStyle(
-                              color: Theme.of(context).primaryColor.withOpacity(0.7),
+                              color: Theme.of(context).primaryColor.withValues(alpha: 0.7),
                               fontSize: 14,
                             ),
                           ),
