@@ -6,7 +6,7 @@ import 'package:extended_text/extended_text.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:image_gallery_saver/image_gallery_saver.dart';
+import 'package:gal/gal.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider/path_provider.dart';
 import 'package:piwigo_ng/models/album_model.dart';
@@ -302,10 +302,7 @@ Future<XFile?> downloadImage(
       },
       outputPath: localPath,
     );
-    await ImageGallerySaver.saveFile(
-      localPath,
-      name: image.name,
-    );
+    await Gal.putImage(localPath);
     return XFile(localPath);
   } on DioException catch (e) {
     debugPrint("Download images: ${e.message}");
