@@ -7,7 +7,7 @@ import 'package:dio/dio.dart';
 import 'package:dio_cookie_manager/dio_cookie_manager.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:heic_to_jpg/heic_to_jpg.dart';
+import 'package:heif_converter/heif_converter.dart';
 import 'package:piwigo_ng/models/album_model.dart';
 import 'package:piwigo_ng/models/status_model.dart';
 import 'package:piwigo_ng/network/api_client.dart';
@@ -91,13 +91,11 @@ class AutoUploadManager {
     // Remove folders and links
     List<File> files = dirFiles.where((file) => file is File).map<File>((e) => e as File).toList();
 
-    // Convert .heic files to .jpg
+    // Convert heic/heif files to .jpg
     for (File file in files) {
-      if (file.path.endsWith('.heic')) {
-        debugPrint("${file.path} is Heic ! Converting to jpg... ");
-        String? jpgPath = await HeicToJpg.convert(
-          file.path,
-        );
+      if (file.path.endsWith('.heic') || file.path.endsWith('.heif')) {
+        debugPrint("${file.path} is Heic/Heif ! Converting to jpg... ");
+        String? jpgPath = await HeifConverter.convert(file.path, format: 'jpg');
         if (jpgPath != null) {
           files.remove(file);
           files.add(File(jpgPath));
